[Generic] support c.OAuthenticator.openid_prefer_ux_over_csrf
This allows the administrator to weigh the consequences of a successful
CSRF attack against the user experience of logging in when using an
OpenID login-by-email provider (eg. Portier[1]).

This should not be done on an OAuth2 provider as an access_token may
be leaked, unless another protection mechanism such as PKCE is used.

[1] https://portier.github.io/